backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
PyMuPDF==1.24.0
Pillow==10.4.0
pydantic==2.9.0
